CLI
Command line interface implemented.

The cli command pipes an event handler to stdin and reads a line, parses it as an array of integers, runs coutCastles function over the array, and, finally, displays the results for the user.

If user types 'quit', then, quits the program.

If an unexpected input is passed or any unexpected behavior occur, display a funny message to the user.

Signed-off-by: Geraldo B. Landre <geraldo.landre@gmail.com>