Documentation & Refactoring
countPeaks and countValleys now receives the array of directions
as parameter instead of the actual array of heights. This saves time and space. Test set updated to consider that.

Comments are improved, and also an explanation about the complexity of the algorithm was added.

Finally, the readme was updated, containing the Solution Approach and the Complexity explanation.

Signed-off-by: Geraldo B. Landre <geraldo.landre@gmail.com>